+++ ChangeLog.md
@@ -1,7 +1,13 @@
 Version 1.0.1-dev
 -----------------
 - Added `hasel` debug feature.
-- Added `--debug-no-abstract-csr` command-line option.
+- Added `--dm-no-abstract-csr` command-line option.
+- Renamed `--progsize` to `--dm-progsize`.
+- Renamed `--debug-sba` to `--dm-sba`.
+- Renamed `--debug-auth` to `--dm-auth`.
+- Renamed `--abstract-rti` to `--dm-abstract-rti`.
+- Renamed `--without-hasel` to `--dm-no-hasel`.
+- Added `--dm-no-halt-groups` command line option.
 
 Version 1.0.0 (2019-03-30)
 --------------------------